backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai
python-dotenv>=1.0.0
jsonschema>=4.0.0
python-docx>=1.0.0
reportlab>=4.0.0
